Only collect app regions for Window Controls Overlay & Borderless apps
This change improves performance for sites that use the CSS property
`app-region` by choosing to not collect the draggable and
non-draggable regions defined by this property when the site is opened
in a normal browser tab since a normal browser tab does not make use
of draggable/non-draggable regions.

By default, `LocalFrame::SupportsAppRegion()` is `false`, and it is
only set to `true` for PWA windows with either Borderless or Window
Controls Overlay enabled.

To support app regions in PWA windows, we check
`AppBrowserController::OnTabInserted/Removed` to handle the case where
web content is reparented into or out of an app window.

To support deprecated Chrome Apps (which are supported on ChromeOS
until Jan 2025), set SupportsAppRegion to true for frameless apps
in `NativeAppWindowViews::RenderFrameCreated`.

Bug: 1447586
Change-Id: Ia96e0bb702e3e9207cf1d145c8d233cc4ce6f3de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4814003
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Amanda Baker <ambake@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1236783}
NOKEYCHECK=True
GitOrigin-RevId: 77422c1c200642af30477fba2379812e7cf75b9c