[Blink] DrawingBuffer: Set ColorSpace after creating SI
With Mappable SI, DrawingBuffer won't have the ability to set the
ColorSpace on the IOSurface until after the SharedImage has been
created. This CL makes that ordering flip in advance to smooth the port.

Reasoning from ccameron@ on why the ordering of the two calls doesn't
matter here:

"The reason for the SetColorSpace() call is that CoreAnimation uses the
IOSurface's color space, while viz uses the SharedImage's color space.
Ideally they should be the same.

If we call CreateSharedImage with a GMB, then we assume that the
IOSurface is having its color space set somewhere else, and we don't set
it [on the service side]. This introduces the possibility that the
IOSurface and the SharedImage might have different color spaces, which
was the bug that resulted in that call being made.

With that background: The order shouldn't matter at all, rather, all
that matters is that both the IOSurface and the SharedImage are
eventually set to have the same color space."

Bug: 1431497
Change-Id: I3fa4ffed61fa12dc68fe030be0d370895e3c6105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5110802
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1236836}
NOKEYCHECK=True
GitOrigin-RevId: a5b552ce5bbff4c96a63563956e04d33dadfe51e