[SharedImage] Convert Blink usage of SHARED_IMAGE_USAGE_GLES2
We're getting rid of SHARED_IMAGE_USAGE_GLES2 in favor of
SHARED_IMAGE_USAGE_GLES2_READ and SHARED_IMAGE_USAGE_GLES2_WRITE. The
former is currently a constant that is equal to the bitwise-or of the
two latter. This CL replaces usage of the former with the bitwised-or
of the latter in //third_party/blink.

Bug: 1510440
Change-Id: Ib8a45a295af1b5deeaf07695781da1b4e53d3320
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5116182
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1236867}
NOKEYCHECK=True
GitOrigin-RevId: ea239cb04a682e5e76fe46178f3f341a28cf374d