Split out context information from BackgroundImageGeometry
This moves the functionality that depends on the box/fragment/other
entity that the image geometry is computed for. The context class is
named BoxBackgroundPaintContext.

This is an as verbatim code movement operation as possible given the
circumstances. In a future CL the context object will be completely
separated from BackgroundImageGeometry, and users will be changed to
create an instances of BoxBackgroundPaintContext instead.

Bug: 1500572
Change-Id: I7a2478b064a88a4279ab0bd35c1e72d49eeb2a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5094017
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1236872}
NOKEYCHECK=True
GitOrigin-RevId: 8dba815ad87d8408bdc3579be1abe7a157bd599f