Refactor blind_sign_auth_proto to use proto deps
Some context: A proto_library works by generating two different targets, let's call the main target "A"

"A" is a static library that will depend on the output of an action. This action is called "A_gen" which takes protobuf
files and generate cpp files.

When we add an entry to `deps` in a `proto_library`, this dependency is added twice, once for target "A"
and another for target "A_gen". Adding it twice is redundant in this case because it is not desirable to
link `anonymous_tokens_cc_proto` because it is being pulled from another path as seen below also this
follows the structure that is recommended by crbug/938011.

//net:net_unittests --[private]-->
//net:net_unittests__library --[private]-->
//third_party/anonymous_tokens:anonymous_tokens_tests --[private]-->
//third_party/anonymous_tokens:anonymous_tokens_pb_openssl_converters --[public]-->
//third_party/anonymous_tokens:anonymous_tokens_cc_proto

In an ideal world, there should be a way to only depend on a fileset that includes
the proto files instead of depending on an action that transforms proto file to cpp files but
unfortunately, GN does not have any support for fileset.

Using `import_dirs` is unrecommended in Chromium because it circumvents proto checkdeps which lead
to using dependencies that might be there due to sheer luck.

Bug: 938011
Change-Id: I1d641e1075d7c0e2c123dc87c7f1b7efe93f9b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5088887
Auto-Submit: Mohannad Farrag <aymanm@google.com>
Commit-Queue: Mohannad Farrag <aymanm@google.com>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1236909}
NOKEYCHECK=True
GitOrigin-RevId: 176a096b1a84e3086db1e62c4ee7008dfefc2fed