Add test support for color providers in Blink::Page
This CL adds test support for the color provders that exist in
Blink::Page. The color providers were introduced in this CL:4898493.
Color providers will be used for painting and as such it is important
to make sure they are not null even in tests.

The CL implements a MockColorProviderSource class in render_view_test.cc,
which simulates the color provider source that is normally provided by
the browser process. The GetRendererColorMap method is used to set the
color_provider_colors view param, which is used by the renderer process
to create the color providers.

The CL also adds a utility function, CreateColorProviderForBlinkTests,
in color_provider_utils.h, which helps create color providers in
blink tests that do not use IPC. The function is called in
DummyPageHolder() and frame_test_helpers::WebViewHelper to ensure that
the color providers are not null even in tests.

Bug: 1430181
Change-Id: I67fc27bdaf58c1a8d5557e1a4158ebdb58dad46a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5046572
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237014}
NOKEYCHECK=True
GitOrigin-RevId: 782a4d32456289600f1781797a74778e5ac9adfc