Rename URLPattern Component HasRegexpGroups to HasRegExpGroups.
To make the function name consistent with URLPattern, let me rename the
function name from HasRegexpGroups to HasRegExpGroups.

I did not touch the liburlpattern::Pattern method name because I think
it natural for the separated library to have a different naming rule.
Also, there is already `GenerateRegexString()` it might be natural to
keep the name aligned with it.

Change-Id: I71dd2300342055f682d55cc476c9f3bc19db00f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119070
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237260}
NOKEYCHECK=True
GitOrigin-RevId: 3423ee1494a64d310ce0ef9b2756e46883e9bd64