Always send mouse up event to main frame during autoscroll selection
We usually route mouse event to the renderer frame that painted the
part of screen that the mouse cursor is currently over. However, there
is cases where mouse down started in main frame and user drag over to
the OOP child frame. The mouse up event was routed to the child frame.
However, the main frame still has autoscroll selection started and is
waiting for next mouse event. Since the mouse event was delivered to
the child frame, this makes the main frame stuck in the autoscroll
state.

Fix is to send a message from main render frame to the browser process
input routing code when there is ongoing autoscroll selection. We will
dispatch mouse up events to both the OOP child frame and main frame
when the state is active.

Bug: 1201368
Change-Id: I53116ebbae39faf254b2c5cfc451321252d61ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5067115
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Siye Liu <siliu@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237380}
NOKEYCHECK=True
GitOrigin-RevId: ebc68b24fcba5e3aaead5a488ed78f4359c43b30