[IntersectionOptimization] Fix crash with fixed-position when viewport is scrolled
A fixed-position element that is fixed to the viewport doesn't scroll
with the viewport, so if the root is the containing LayoutView, we
don't need to update intersection on viewport scroll. However,
previously CachedRects::unscrolled_unclipped_intersection_rect was
always in the scrolling contents coordinates making it depend on
viewport scroll position, and caused crashes when
IntersectionOptimization skipped intersection update on viewport
scroll.

Now let CachedRects::unscrolled_unclipped_intersection_rect be in the
root's border box space (i.e. viewport coordinates if the root is the
viewport) if the root doesn't scroll the target.

Bug: 1508301
Change-Id: I94fa8e60610ceae2826cc560aa99f568d67be583
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5106931
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237621}
NOKEYCHECK=True
GitOrigin-RevId: 75bd1751bf01d0054b24b1ec495e944f04ef872c