Remove CHECK for missing cached cookie
The point of the check was to validate what was assumed an invariant. Crash
reports are coming in at a rate that is high for crashes but extremely
low in the context of the number of cookie operations. This is caused
by the mojo operation returning a null string which can happen.

If a null value is detected the code now falls back on using an IPC to request a new value.

Bug: 1393050, 1511021
Change-Id: I0bc52eed3799b95d73ec1f1c4c01665356b11f31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119481
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237641}
NOKEYCHECK=True
GitOrigin-RevId: 4178e369978d840981cbb446bb5cb861858f9ad2