Do not add kBusy to popup root
When adding a popup to the accessibility tree, the root
popup node is a Document node but a "group" role. The
busy attribute should not be put on this node, only on the
root document node.

This was discovered in process of working on a flaky test:
https://chromium-review.googlesource.com/c/chromium/src/+/5050394

Bug: 1055764
Change-Id: Idc9a3ab248b128cbfedcec68cb24b89984647003
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5105159
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Joanmarie Diggs <jdiggs@igalia.com>
Commit-Queue: Valerie Young <spectranaut@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1237650}
NOKEYCHECK=True
GitOrigin-RevId: fee1763367432955bfae136a2f85a8a661660278