Update resetInternal to explicitly reset recording
BaseRenderingContext2D::ResetInternal wasn't explicitly resetting the
recording. It was depending on a side effect of calling clearRect,
which has an overdraw optimization clearing the recording if the rect
extents covers the whole canvas. To make this work however,
ResetInternal has to hack into the canvas, manually resetting it to
its default state, or else, any transform or clip would prevent the
overdraw optimization to run.

Moreover, the overdraw optimization can actually be disabled via the
kDisableCanvasOverdrawOptimization feature. Doing so would actually
break ResetInternal, preventing it from dropping the previous recording.

Apart from this, the manual reset of the canvas is an abuse of
responsibility and an encapsulation problem. A follow up CL will split
the recording in two parts and this code would need to be updated to
correctly maintain the split recording state. It's much cleaner to
delegate that logic to the right abstraction.

Bug: 1484741
Change-Id: Ica3257f8ad38b98b0c9e9346697d0d229f2b40c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4953275
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237740}
NOKEYCHECK=True
GitOrigin-RevId: 9bf69d1380b61f61b62167890579bd7a06909ad3