[soft navigations] Remove ScriptState from SoftNavigationEventScope
The ScriptState was ultimately being used for ResetPaintsIfNeeded() in
order to get the frame (via ToLocalFrameIfNotDetached) -- but this
isn't necessary. SoftNavigationHeuristics is a supplement of
LocalDOMWindow -- so we can get the frame from the supplementable and
check whatever properties are needed. This avoids needing to create
any v8 locals when we just need to query dom object properties.

Bug: 1351643
Change-Id: I70bf36a303cbd2526e17ee40b8d7e2f74d1cbaf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5118760
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237762}
NOKEYCHECK=True
GitOrigin-RevId: ce05c1eeadd3d87ece32c969a5776d23f224d70d