[a11y] NextOnLine and PreviousOnLine not working with inert elements
A bug was found where AXObject::NextOnLine/PreviousOnLine were not
working when we had inert elements in the line.

So on a structure that looks like this:

<span id="span1">this </span><span inert>is</span><span> broken</span>

* "span1"'s NextOnLine would return nullptr.
* The last span's PreviousOnLine would return nullptr.

This CL fixes this bug by making sure we continue to move the cursor
until we find the next/previous non-inert in line element or until
we can't move the cursor anymore.

Change-Id: I900692aae3c538835960a15258b1f4708d918038
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5118351
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1237858}
NOKEYCHECK=True
GitOrigin-RevId: 89f66ae93f4b7a1fef4a365e16167ab16298c4db