Pass PendingRemote<CodeCacheHost> for background on CommitNavigation
Currently one pending_remote<blink.mojom.CodeCacheHost> is passed on
the CommitNavigation IPC (https://crrev.com/3033191). It is used only
for normal resource fetch from the main thread. When loading a resource
from the background thread, LoaderFactoryForFrame::
GetBackgroundCodeCacheHost() gets a new pending_remote of
blink.mojom.CodeCacheHost from BrowserInterfaceBroker. This is not good
for performance, because it can be blocked if the browser main thread
is busy.

This CL solves this problem by sending two pending_remotes of
CodeCacheHost on the CommitNavigation IPC, for the normal
resource fetch from the main thread and for the resource fetch from the
background thread.

Bug: 1379780
Change-Id: I4267083ab6c50923e7c1d0d753278d0a65135c39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5114340
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237952}
NOKEYCHECK=True
GitOrigin-RevId: 4d335dfbdb52a7bac7cd3bc2644000a9893676e7