[A11y] No calls to GetOrCreate() unless processing deferred events
- Assert that GetOrCreate() is only called while processing
  deferred events.
- At other times, e.g. when the tree is frozen, we should just be using
  Get(), because the tree is already complete, and because it is unsafe
  to modify the tree structure at that time anyway.

This helps prepare for CL:5020124, where we will remove GetOrCreate()
calls that only take a node (a parent will be required). That signature
will be removed because AXObjects will only be created when adding them
from the top down. Ultimately, this helps move toward the goal of
never needing to "repair" a tree with missing parents.

A follow-up will be to replace calls to SafeGet() with Get(), since
they now do the same thing.

Bug: none
Change-Id: I21c473ed5a0f3420111bafcf163c8a0ec576b147
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5123413
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1238297}
NOKEYCHECK=True
GitOrigin-RevId: d583460a04196a7f296350ae493bde0b80b3ccd4