Make edge carets for atomic inline blocks painted by their container
In https://crrev.com/c/3620254, it became possible to have carets
before or after a non-editable inline-level element, if there was
nowhere else to place the caret without crossing into the non-editable
element. Before this, all carets inside an inline formatting context
would have been placed at a text offset, rather than at the edges of
an element.

When `CaretDisplayItemClient` computes the caret rectangle for a caret
position, it also computes the containing block that should paint it,
and it does that by using the caret's layout object if it's a
`LayoutBlock`, or its containing block if it isn't. For carets at the
edges of non-editable atomic inline blocks (e.g. inline-block,
inline-flex...), this gets the wrong result, since the layout object
is a `LayoutBlock`, but the caret should be painted by the inline
formatting context's containing block. In those cases, the caret ends
up not being painted.

The method in `CaretDisplayItemClient` that computes the painter block
doesn't know whether the caret is at the edge or inside an atomic
inline block, and so it can't know whether to use its containing
block. However, `LocalCaretRect` also contains a root box fragment,
whose layout object seems to match the computed painter block, except
that for carets at the edges of atomic inline blocks where it is the
inline formatting context's container.

This patch changes the code that computes the painter block to use
the root box fragment's layout object, if the root box fragment is
non-null, and uses the old algorithm to compute the painter block
otherwise. It also adds DCHECKs to make sure the behavior doesn't
change for cases other than atomic inline blocks, and to ensure that
the invariant of the painter block being either the caret's layout
block or its direct container keeps being held.

Bug: 1499405
Change-Id: Id552035b56a7bd51c4f90029801e9d3f52bb765f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5110302
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1238390}
NOKEYCHECK=True
GitOrigin-RevId: 44dd9f657b810272e9b6a53f58baf153c2b02957