[bfcache] Allow BFCache where there's ended MediaStreamTrack caused by getDisplayMedia()
This CL basically makes pages with no live Media Stream Track eligible for BFCache. When there's an "ended" (so-called "stopped", "closed") Media Stream Track caused by the use of the getDisplayMedia() method of MediaDevicesAPI, the page will be stored in BFCache. Note that the page will NOT be stored in BFCache when there's a "live" (so-called "active", "open") Media Stream Track.

The exception is that when there's an ended Media Stream Track caused by getUserMedia(), not getDisplayMedia(), the page will still be blocked from BFCaching for the reason kRequested(Audio/Video)CapturePermission.

The feature flag to guard this new implementation is defined in third_party/blink/common/features.cc because it needs to be used from both the blink and content components.

Also, this CL adds a new test to back_forward_cache_features_browsertests.cc, which checks that a page will be BFCached after a Media Stream Track is ended when getDisplayMedia() is used.

Bug: 1502395, 989379
Change-Id: I46bc6d960b3ffe6bbda1255e8d488143cb873413
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5075713
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Kurumi Muto <mkurumi@google.com>
Cr-Commit-Position: refs/heads/main@{#1238506}
NOKEYCHECK=True
GitOrigin-RevId: ded8f577facd8293eab73a524c2d04d32759bf41