[Text Contrast] Move SKSurfaceProps usage to default constructor
I am planning on adding a new constructor to SKSurfaceProps in a follow
up CL that will add text contrast and gamma parameters. The end result
of this will be two main constructors for SKSurfaceProps:

1) Default/empty constructor with all default values
2) A populated constructor with hardware/system values (pixel geometry,
contrast, gamma, flags, etc.)

There is an existing pattern of mixing these concepts by calling
into a constructor that allows for non-default values, but manually
passing in the defaults, which this CL removes. There should be no
behavioral changes as a result of this CL.

The corresponding Skia change is here:
https://skia-review.googlesource.com/c/skia/+/789118

Bug: 1160653
Change-Id: I883f28db96c465f204ed8fb42f596454718b8dbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5118781
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1238669}
NOKEYCHECK=True
GitOrigin-RevId: 0ca4d29db3b3299841a89b829d3e8a09050ead2f