Remove extra LCP Candidate Update
The LCP candidate update should be part of the presentation callback
registered by Image/TextPaintTimingDetector. Currently it is
missing in the TextPaintTimingDetector callback and is made up by
having one after the invocation of all callbacks. This works but
is logically unclear.

This CL is to fix this.

Change-Id: Ia476a17b25ea8bc6942f3670438a13ed89aa5acb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5057910
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239185}
NOKEYCHECK=True
GitOrigin-RevId: bd74926736a9518f485d5fd237422abb58d8edde