[RWS] Remove kStorageAccessAPI feature
This feature has been enabled by default, with the default values for
all of its FeatureParams, since https://crrev.com/c/4874417.

This removal is mostly mechanical, with some exceptions for unit- and
browsertests: many tests actually rely on the ability to prompt
the user for permission, or assume that prompts are disabled; but
this is controlled by the `kPermissionStorageAccessAPI` feature, which
is not appropriately controlled for these tests. I added test fixtures and migrated all of the tests that make such assumptions.

I also preserved the "implicit grants" functionality and tests, even
though Chrome does not use it, because Edge sets the implicit grant
limit to a nonzero value (5, IIRC).

Change-Id: I9a69dee1200a064d1bbb230eb6f09e2e97dfa0ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124980
Reviewed-by: Florian Jacky <fjacky@chromium.org>
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239255}
NOKEYCHECK=True
GitOrigin-RevId: d1ed70df6bb14e803288b9036dd897471951c081