[A11y] When unpausing an update, rebuild tree immediately
When tree updates are paused, we mark the document dirty so that it
is rebuilt. This CL changes the code so that we rebuild the
tree immediately when unpausing, rather than leaving the tree
in an incomplete state. This fixes the referenced Clusterfuzz bug where  objects were accessed at a time that object properties were expected to
be frozen, but the properties were marked dirty. A unit test is added
to simulate the conditions created by the Clusterfuzz test.

Note: when tree updates are paused, it's because there are too many updates, and we make a guess that just rebuilding the tree from the root
is the better option for performance, even though this could cause the
user to lose their place because all of the newly generated AXObjects
will have new AXIDs. This problem will be helped by stable ids
(CL:4185636).

Bug: 1512428,1507170
Change-Id: I475ad5af5010ca4d273e54c4ac4865a0d89e68bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5089911
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239260}
NOKEYCHECK=True
GitOrigin-RevId: a9c2ae7f54c7067c22141d44172c6f620b3278b8