Fragmentation: Simplify gathering of OOF descendants to lay out.
If we simply allow OOF descendant propagation when cloning the old
fragments (instead of disabling it), we get this list for free. This way
we won't have to look at the old physical fragment and convert
everything from physical to logical.

We now grab the OOF fragmentainer descendants off the builder, so that
they don't get repropagated further up (where they don't belong; we're
peeking inside a nested multicol at this point). We are not interested
in the OOF "candidates", but we similarly need to make sure that they
don't get repropagated, so this list now needs to be manually cleared.

No behavior changes intended.

Change-Id: I5fbef59b0cac2d8cc3e2d7827186b754300581fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5134888
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1239310}
NOKEYCHECK=True
GitOrigin-RevId: c982ba64db0ba6c12aaf808bc96283f0876eb466