media: Add Media.EME.{KeySystem}.{Robustness}.WaitingForKey UMA
This is to monitor cases where sites which didn't properly set a CDM or
fetch a license before an encrypted media playback starts. This could
lead to time-to-play delay or player stuck with playing.

In a later CL we'll add a UKM as well so we would understand which sites
are experiencing this.

Manually tested with playbacks and saw
```
- Histogram: Media.EME.Widevine.SoftwareSecure.WaitingForKey recorded 4 samples, mean = 0.0 (flags = 0x41) [#]
0  ----O                                        (4 = 100.0%)
1  ...
```

Also hacked Chromium to not pass the license to the CDM, and saw
```
- Histogram: Media.EME.Widevine.SoftwareSecure.WaitingForKey recorded 1 samples, mean = 1.0 (flags = 0x41) [#]
0  O                                           (0 = 0.0%)
1  -O                                          (1 = 100.0%) {0.0%}
2  O                                           (0 = 0.0%) {100.0%}
```

Bug: b/281532789
Test: Manually tested. See above.
Change-Id: I1e8b757693bff22383823d2b6d7a25e495ff30bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5133770
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239369}
NOKEYCHECK=True
GitOrigin-RevId: aeea471eaa248f15caa845325eeb339fd8f4eb60