[@position-fallback] Make sure LayoutObject style is correct
When trying to make room for absolute positioned with position-fallback
the style without animations applied is used. Since TryCalculateOffset
uses a mix of the passed in style and the style retrieved from the
LayoutObject, we need to set the base style on the LayoutObject to make
sure they are in sync.

Bug: 1512373
Change-Id: Id7bc35b11a5c9b2b7c0ac20883ff457f4d8bbfd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5134895
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239412}
NOKEYCHECK=True
GitOrigin-RevId: d5f2ecde83223022795eff3c8a92527fb9303287