Compute ink overflow for all the document markers
Document marker ink overflow loops over the markers for the node to
which a fragment belongs, with a check to ignore markers outside the
fragment range. That check was returning from the function on the first
out-of-range marker, failing to compute overflow for subsequent markers.
Fix it by continuing the loop rather than returning.

Add a test to cover this situation.

Bug: 1475771
Change-Id: I43b753c9e1daf4ab44f848ce6669258763575204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5139906
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1239684}
NOKEYCHECK=True
GitOrigin-RevId: abb229274caeb693ff3dff7ebbf165084df94cfc