Fix libxml TLS configure for windows
When I added the --without-tls option to the libxml configure scripts in
crrev.com/1235361 it caused the configure script system for libxml on
windows to stop working because that configure script doesn't implement
TLS options.

This patch stops passing that argument to the windows configure script.
I also ran the configure script on mac which commented out the macro
which enables TLS on mac. I feel confident that TLS is not currently
enabled on windows because the generated windows header does not define
the macro which enables TLS.

Bug: 1508362
Change-Id: Ia5c5f973670715e566ab411a037b40a64b791472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5120129
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239702}
NOKEYCHECK=True
GitOrigin-RevId: 8a43c14e277e28f9aa1c45282314b8074c89dafb