Notify the flow thread about style change for inlines as well.
We only did it for LayoutBox or better, but this is necessary for
LayoutInline as well.

When a layout object changes whether or not it's a containing block for
out-of-flow positioned descendants, the multicol flow thread machinery
may need to insert or remove LayoutMultiColumnSet objects - since
ouf-of-flow positioned descendants may become part of (or cease to be
part of) the multicol container because such a change.

This is detected in NeedsToReinsertIntoFlowThread().

Move the flow thread hooks in StyleWillChange() and StyleDidChange()
from LayoutBox to LayoutBoxModelObject, so that this also works for
inlines.

This was most likely a LayoutNG regression, since the legacy engine
didn't really let inlines act as containing blocks.

Bug: 1503298
Change-Id: I7afcf447f8d81bd35cb1f049d04c7cdae3d3abf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5137017
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239713}
NOKEYCHECK=True
GitOrigin-RevId: 234e35eb3ae4fa77f193065dd35e7f083f71077b