[A11y] Partial revert to fix crash when finding <option> parent
Partial revert of CL:5123413. Remove NOTREACHED() and restore comment.

This condition is rare, but can occur if a <select> is inside of a
<details> element. When the <select> is first created, there is
no layout object, and a generic AXObject (not AXMenuList) is
created. The autofill code can call in just at the right time to
request this parent, and hits the NOTREACHED().

Test: this crash occurs when loading clock.laplace.live, or by loading the included test in the browser. It's unclear why, but the test does
not crash when running as a web test.

Note that these AXMenuList* classes will be going away in favor
of a generic implementation based on the shadow DOM, using ordinary
AXNodeObjects.

Fixed: 1512999
Change-Id: I318d6198abc7b155e42030caf00f039c7b09e4ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5141647
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1239844}
NOKEYCHECK=True
GitOrigin-RevId: deb04292809448bad1cf756536f6a411fa2a215b