Fenced frame: Add ad component remote context WPT support.
Introduces a way to embed ad components within frames and be able to run
arbitrary JavaScript in them.

Steps for setting up:
1. Create a fenced frame or URN iframe remote context using either
attachFencedFrameContext() or attachIFrameContext(). Specify the new
`include_component: true` parameter when calling the function. This
will create a remote context URL/URN, but not actually create a frame.
2. To create the ad component frame, call either
attachComponentFencedFrameContext() or attachComponentIFrameContext().
This will construct an ad component frame, and return the remote context
object the same way that attachFencedFrameContext/attachIFrameContext
does. This can be interacted with in the same way that any other remote
context frame can be interacted with.

As a proof of concept, this CL makes the following test changes:
1. Updates `nested-opaque-ad-sizes.https.html` to finally be functional.
2. Adds a WPT to test that ad components can send automatic beacons with
a header opt-in.

Change-Id: Ic4277b4082e36617f95f20895c9502bca3928965
Bug: 1427455
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5136743
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1240220}
NOKEYCHECK=True
GitOrigin-RevId: c611a85b90716131fe8d9e3a8df3265fb2e65a1f