Add a screen-ai CIPD package for PDF OCR E2E tests on Linux
To have E2E tests for PDF OCR on Windows, macOS, and Linux, we need to
be able to download and use the screen-ai library in those tests.
Pursuing this goal, we upload the screen-ai library for Linux to CIPD
and download it from CIPD to //third_party/screen-ai. This downloaded
library is then loaded in E2E tests (i.e. browser tests).

For more information, visit go/chrome_screen_ai_tests and read the
Uploading the library files to CIPD section.

AX-Relnotes: n/a.
Bug: 1443346
Change-Id: I7d55f9ae970a8e39aafbc4273fae2d654b4af17e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4803348
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240288}
NOKEYCHECK=True
GitOrigin-RevId: 8c005672e75f619a1f2df8241fe996f9f03cbe7d