Move metric lcp update to lcp calculator
To move the lcp update for reporting from PaintTimingDetector to
LargestContentfulPaintCalculator.

The lcp_object.h is used to put in lcp related classes that are used in
multiple places. It's easier to avoid circular dependencies.

This is part of the LCP code refactor. See refactor design
https://docs.google.com/document/d/1VNt5VTNYkvke6dHg6Y95-pKWkqPKrJ-xKiD6e4BaFRs/edit?usp=sharing

Change-Id: Idc6183928fcb3aa1efdb571b3d9f8fdf0c6d457b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5067721
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240314}
NOKEYCHECK=True
GitOrigin-RevId: bbc1bf3c26629c97a1b9654c9da5e6c67ae5764c