Fix clipboard crash when reading files on Lacros.
The root problem is that multiple IPC clipboard calls are made in a row
and the actual clipboard state can change while that's happening, which
leads to a situation where the clipboard may fail to serve up types it
recently claimed to have had on offer.

This race should not in theory be limited to Lacros, but in practice I
suspect Lacros is doing something weird with the clipboard such as
modifying it in the background without user intervention.

The proximate cause is not checking the mojo smart pointer type for
nullness, which is how this patch addresses the issue. The paste will
still fail, but silently, which may be something for the CrOS team to
investigate.

Bug: 1512549
Change-Id: I6982487919b9e90d1f71bd7cb6cf6b6c406ca3e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5146848
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240325}
NOKEYCHECK=True
GitOrigin-RevId: 77dbe7b468346dae310cf5e03ee74fb1de8dba38