Disallow invalid node types in DOM Parts API constructors
It was previously possible to construct, e.g., a ChildNodePart with
previousSibling and nextSibling being the DOCTYPE element and the
documentElement. That causes problems when trying to clone that part,
because the CloneParts code tries to cast the ChildNodePart to a
DocumentPart, which it isn't.

This CL also adds more testing for this set of cases, and sets a few
more of the `DCHECK`s to `CHECK`s. This might be a performance hit on
the perf tests, but I think it's worth it.

Fixed: 1510578
Change-Id: Ie18901857a4b089265185c3efdabfb5448da2e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5142457
Commit-Queue: Dan Clark <daniec@microsoft.com>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1240329}
NOKEYCHECK=True
GitOrigin-RevId: bb7c52dcbbac1c5e09cebab2b856816a28c07756